feat: Code refactoring and Swagger update
