+++ api/v1/app.ts
@@ -6,6 +6,8 @@ import swaggerJSDoc from 'swagger-jsdoc';
 import AnalysisRouter from './modules/analysis';
 import UserRouter from './modules/user';
 import HistoryRouter from './modules/history';
+import ProfessionalExperienceRouter from './modules/professional-experience';
+
 import cors from 'cors';
 
 const app = express();
@@ -34,6 +36,7 @@ app.use('/v1/docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 app.use(AnalysisRouter);
 app.use(UserRouter);
 app.use(HistoryRouter)
+app.use(ProfessionalExperienceRouter)
 // Server -------------------------------------------------------
 connectToMongoDB()
   .then(() => {
