+++ api/v1/modules/professional-experience/index.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import ProfessionalExperience from './routes';
+
+const router = express.Router();
+
+router.use('/professional-experience', ProfessionalExperience);
+
+export default router;
