+++ api/v1/modules/user/routes.ts
@@ -3,16 +3,13 @@ import express from 'express';
 import {
   getAllUser,
   getUserById,
+  getProfessionalExperiencesByUserId,
   createCandidate,
   createRepresentative,
   updateCandidate,
   updateRepresentative,
   deleteUser,
-  loginUser,
-  getProfessionalExperiencesByUserId,
-  createProfessionalExperience,
-  updateProfessionalExperience,
-  deleteProfessionalExperience
+  loginUser
 } from './controllers/UserController';
 import {
   checkCreateCandidate,
@@ -20,9 +17,6 @@ import {
   checkUpdateCandidate,
   checkUpdateRepresentative,
   checkLoginUser,
-  checkCreateProfessionalExperience,
-  checkUpdateProfessionalExperience,
-  checkDeleteProfessionalExperience,
   checkGetUserById,
   checkGetProfessionalExperienceByUserId, 
   checkDeleteUser} from './middlewares/UserMiddleware';
@@ -34,15 +28,12 @@ const router = express.Router();
 // Revisar rutas de experience(experiences o experience)
 router.get('/', getAllUser);//X
 router.get('/:id', checkGetUserById, getUserById);//X
-router.get('/:id/experiences', checkGetProfessionalExperienceByUserId, getProfessionalExperiencesByUserId);//X
+router.get('/:id/professional-experiences', checkGetProfessionalExperienceByUserId, getProfessionalExperiencesByUserId);//X
 router.post('/candidate', checkCreateCandidate, createCandidate);//X
 router.post('/representative', checkCreateRepresentative, createRepresentative);//X
 router.post('/login', checkLoginUser, loginUser);//X
-router.post('/experience', checkCreateProfessionalExperience, createProfessionalExperience);//X
 router.patch('/candidate/:id', checkUpdateCandidate, updateCandidate);//X
 router.patch('/representative/:id', checkUpdateRepresentative, updateRepresentative);//X
-router.patch('/experience/:id', checkUpdateProfessionalExperience, updateProfessionalExperience);//X
 router.delete('/:id', checkDeleteUser, deleteUser);//X
-router.delete('/experience/:id', checkDeleteProfessionalExperience, deleteProfessionalExperience);//X
 
 export default router;
